Added printCpuFeatures function to lib.c on kernel

+++ Kernel/include/console.h
@@ -2,6 +2,7 @@
 #define CONSOLE_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 typedef enum Color {
 	Black, Blue, Green, Cyan, Red, Purple, Brown, Gray, DarkGray, LightBlue, LightGreen, LightCyan, 
@@ -16,8 +17,8 @@ void println(char *string);
 void printlnCol(char *string, uint8_t foreColor, uint8_t backColor);
 void printChar(char character);
 void printCharCol(char character, uint8_t foreColor, uint8_t backColor);
-void printInt(int num, uint8_t base);
-void printIntCol(int num, uint8_t base, uint8_t foreColor, uint8_t backColor);
+void printInt(size_t num, uint8_t base);
+void printIntCol(size_t num, uint8_t base, uint8_t foreColor, uint8_t backColor);
 void printErr(char *string);
 void newLine();
 void clearScreen();
